web_server/requirements.txt
pika==0.11.2
Django==2.0.1
django-webpack-loader==0.6.0
djangorestframework==3.7.7